0.1.26
git-svn-id: https://plugins.svn.wordpress.org/wp-symposium/trunk@333264 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ ajax/symposium_forum_functions.php
@@ -73,7 +73,9 @@ if ($_POST['action'] == 'updateForum') {
 	$action = $_POST['value'];
 
 	// Store subscription if wanted
-	$wpdb->query("DELETE FROM ".$subs." WHERE uid = ".$current_user->ID." AND tid = ".$tid);
+	if (symposium_safe_param($tid)) {
+		$wpdb->query("DELETE FROM ".$subs." WHERE uid = ".$current_user->ID." AND tid = ".$tid);
+	}
 	
 	if ($action == 1)
 	{		
@@ -148,7 +150,9 @@ if ($_POST['action'] == 'updateForumSubscribe') {
 	$cid = $_POST['cid'];
 
 	// Store subscription if wanted
-	$wpdb->query("DELETE FROM ".$subs." WHERE uid = ".$current_user->ID." AND tid = 0 AND (cid = ".$cid." OR cid = 0)");
+	if (symposium_safe_param($cid)) {
+		$wpdb->query("DELETE FROM ".$subs." WHERE uid = ".$current_user->ID." AND tid = 0 AND (cid = ".$cid." OR cid = 0)");
+	}
 	
 	if ($action == 1)
 	{
